+++ src/backend/commands/admin-command.manager.ts
@@ -3,6 +3,7 @@ import {
     CommonAdminCommandConfigType,
     CommonAdminCommandManager
 } from '@dps/mycms-server-commons/dist/backend-commons/commands/common-admin-command.manager';
+import {ConfigInitializerCommand} from './config-initializer.command';
 
 export interface AdminCommandConfigType extends CommonAdminCommandConfigType {
 }
@@ -10,6 +11,7 @@ export interface AdminCommandConfigType extends CommonAdminCommandConfigType {
 export class AdminCommandManager extends CommonAdminCommandManager<AdminCommandConfigType> {
     constructor(adminCommandConfig: AdminCommandConfigType) {
         super({
+            'initConfig': new ConfigInitializerCommand(),
             'generateSitemap': new SiteMapGeneratorCommand()
         }, adminCommandConfig);
     }